Define Column Widths

function for aiding more personalized column widths in case default is not desired.

Usage

listing_column_widths(
  mpf,
  incl_header = TRUE,
  col_gap = 0.5,
  pg_width_ins = 8.88,
  fontspec = font_spec("Times", 8, 1.2),
  verbose = FALSE
)

def_colwidths(tt, fontspec, label_width_ins, col_gap, type = tlg_type(tt))

Arguments

mpf

(`listing_df` or `MatrixPrintForm` derived thereof)
The listing calculate column widths for.

incl_header

(`logical(1)`)
Should the constraint to not break up individual words be extended to words in the column labels? Defaults to `TRUE`

col_gap

Column gap.

pg_width_ins

(`numeric(1)`)
Number of inches in width for *the portion of the page the listing will be printed to*. Defaults to `8.88` which corresponds to landscape orientation on a standard page after margins.

fontspec

Font specification

verbose

(`logical(1)`)
Should additional information messages be displayed during the calculation of the column widths? Defaults to `FALSE`.

tt

input Tabletree

label_width_ins

Label Width in Inches.

type

Type of the table tree, used to determine column width calculation method.

Value

A vector of column widths suitable to use in `tt_to_tlgrtf` and other exporters.

Examples

# colwidths <- def_colwidths(result, font_spec('Times', 9L, 1.2), col_gap = 7L, label_width_ins = 2)