Helper functions to test proportion differences

Helper functions to implement various tests on the difference between two proportions.

Usage

prop_chisq(tbl, alternative)

prop_cmh(ary, alternative)

prop_fisher(tbl, alternative)

Arguments

tbl

(`matrix`)
matrix with two groups in rows and the binary response (`TRUE`/`FALSE`) in columns.

ary

(`array`, 3 dimensions)
array with two groups in rows, the binary response (`TRUE`/`FALSE`) in columns, and the strata in the third dimension.

Value

A p-value.

Functions

prop_chisq(): Performs Chi-Squared test. Internally calls [stats::prop.test()].

prop_cmh(): Performs stratified Cochran-Mantel-Haenszel test. Internally calls [stats::mantelhaen.test()].

prop_fisher(): Performs the Fisher's exact test. Internally calls [stats::fisher.test()].

Note

strata with less than five observations will result in a warning and possibly incorrect results; strata with less than two observations are automatically discarded.

See also