Split Function Factory for the Response Tables (RESP01)

The main purpose here is to have a column dependent split into either comparative statistic (relative risk or odds ratio with p-value) in the 'Overall' column, and count proportions and corresponding confidence intervals in the other treatment arm columns.

Usage

resp01_split_fun_fct(method = c("rr", "or_logistic", "or_cmh"), conf_level)

Arguments

method

(string)
which method to use for the comparative statistics.

conf_level

(proportion)
confidence level of the interval.

Value

A split function for use in the response table RESP01 and similar ones.

See also

Examples

split_fun <- resp01_split_fun_fct(
  method = "or_cmh",
  conf_level = 0.95
)