Statistical Function for Exposure-Adjusted Incidence Rate

This function calculates exposure-adjusted incidence rates (EAIR) per 100 person-years for a specific level of a variable.

Usage

s_eair100_levii_j(
  levii,
  df,
  .df_row,
  .var,
  .alt_df_full = NULL,
  id = "USUBJID",
  diff = FALSE,
  conf_level = 0.95,
  trt_var = NULL,
  ctrl_grp = NULL,
  cur_trt_grp = NULL,
  inriskdiffcol = FALSE,
  fup_var,
  occ_var,
  occ_dy
)

Arguments

levii

(`string`)
the specific level of the variable to calculate EAIR for.

df

(`data.frame`)
data set containing all analysis variables.

.df_row

(`data.frame`)
data frame across all of the columns for the given row split.

.var

(`string`)
variable name that is passed by `rtables`.

.alt_df_full

(`dataframe`)
alternative dataset for calculations.

id

(`string`)
subject variable name.

diff

(`logical`)
if TRUE, risk difference calculations will be performed.

conf_level

(`proportion`)
confidence level of the interval.

trt_var

(`string`)
treatment variable name.

ctrl_grp

(`string`)
control group value.

cur_trt_grp

(`string`)
current treatment group value.

inriskdiffcol

(`logical`)
flag indicating if the function is called within a risk difference column.

fup_var

(`string`)
follow-up variable name.

occ_var

(`string`)
occurrence variable name.

occ_dy

(`string`)
occurrence day variable name.

Value

A list containing the following statistics:

n_event: Number of events

person_years: Total person-years of follow-up

eair: Exposure-adjusted incidence rate per 100 person-years

eair_diff: Risk difference in EAIR (if diff=TRUE and inriskdiffcol=TRUE)

eair_diff_ci: Confidence interval for the risk difference (if diff=TRUE and inriskdiffcol=TRUE)